fix[redux scope): added new provider in dashboard layout
